pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "outowiki"
version = "0.1.0"
description = "AI agent knowledge management system using Wiki-based markdown storage"
readme = "docs/index.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "OutoWiki Contributors" },
]
dependencies = [
    "pydantic>=2.0",
    "openai>=1.0",
    "anthropic>=0.18",
    "pyyaml>=6.0",
    "mistune>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/outowiki"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py310"
line-length = 100
